feat: refactor stats page and enhance configuration
- Updated Next.js configuration to ignore ESLint and TypeScript build errors.
- Introduced remote image patterns for external sources.
- Refactored the stats page to utilize a new StatsContent component, improving code organization and readability.
- Added functionality for token selection and total value locked calculation in the new StatsContent component.
- Updated package scripts for linting and type checking.
